Finished unit test of CacheSessionService

+++ src/Service/CacheSessionService.php
@@ -9,13 +9,17 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\SimpleCache\CacheInterface;
 use RuntimeException;
 
-// TODO clean up the logic of this class - for now we care about passing tests.
-// TODO more handling of session ttl than cookie lifetime.
+/**
+ * A session service that stores the session data in a PSR-16 compliant cache storage.
+ *
+ * get(), set(), flash(), unset(), and clear() actions are cached individually and stored to the cache at commit().
+ */
 class CacheSessionService implements SessionService
 {
-    const SESSION_COOKIE_KEY    = "kodus-session";
-    const SESSION_INDEX_PREFIX  = "kodus.session.";
-    const FLASHES_STORAGE_INDEX = "kodus.session.flashes";
+    const COOKIE_KEY             = "sessionID";
+    const SESSION_INDEX_PREFIX   = "kodus.session.";
+    const FLASHES_STORAGE_INDEX  = "kodus.session.flashes";
+    const SESSION_EXPIRATION_KEY = "kodus.session.expiration.";
 
     /**
      * @var CacheInterface
@@ -23,32 +27,32 @@ class CacheSessionService implements SessionService
     private $storage;
 
     /**
-     * @var array
+     * @var array Objects to be written to the cache storage on commit()
      */
     private $write_cache = [];
 
     /**
-     * @var array
+     * @var array Level one cache for reading objects from the cache storage.
      */
     private $read_cache = [];
 
     /**
-     * @var array
+     * @var array Type names of objects to be removed from cache storage on commit()
      */
     private $removed = [];
 
     /**
-     * @var bool
+     * @var bool Indicates whether the clear() method was called since the last commit()
      */
     private $cleared = false;
 
     /**
-     * @var array
+     * @var array Index of the object types that was set as flash sessions. These will be removed during next request.
      */
     private $flashed = [];
 
     /**
-     * @var string
+     * @var string The session ID of the current session.
      */
     private $session_id;
 
@@ -57,13 +61,22 @@ class CacheSessionService implements SessionService
      */
     private $session_ttl;
 
+    /**
+     * CacheSessionService constructor.
+     *
+     * @param CacheInterface $storage     The cache provider to store session data in.
+     * @param int            $session_ttl Session lifetime in seconds
+     * @param string|null    $session_id  Session ID string, or null if starting new session
+     */
     public function __construct(CacheInterface $storage, $session_ttl = 60 * 60 * 24 * 14, $session_id = null)
     {
         $this->storage = $storage;
-
-        $this->session_id = $session_id ?: base64_encode(UUID::create());
-
+        $this->session_id = $session_id;
         $this->session_ttl = $session_ttl;
+
+        if (! $this->sessionIsActive($this->session_id)) {
+            $this->session_id = base64_encode(UUID::create());
+        }
     }
 
     public function flash(SessionModel $object)
@@ -111,7 +124,17 @@ class CacheSessionService implements SessionService
     }
 
     /**
-     * Save Session data into storage and add session cookie to response
+     * TODO part of SessionService interface?
+     *
+     * Return the current session ID string
+     */
+    public function getSessionID()
+    {
+        return $this->session_id;
+    }
+
+    /**
+     * Save Session data into the cache storage and add session cookie to response
      *
      * @param ResponseInterface $response
      *
@@ -124,8 +147,14 @@ class CacheSessionService implements SessionService
         }
         $flashes = $this->storage->get(self::FLASHES_STORAGE_INDEX) ?: [];
 
-        foreach ($flashes as $type) {
-            $this->storage->delete($this->storageKeyFromType($type));
+        //If the response is a redirect or error, keep the current flashes in memory, otherwise clean them up.
+        if ($response->getStatusCode() < 300) {
+            foreach ($flashes as $type) {
+                $this->storage->delete($this->storageKeyFromType($type));
+            }
+            $this->storage->set(self::FLASHES_STORAGE_INDEX, $this->flashed);
+        } else {
+            $this->storage->set(self::FLASHES_STORAGE_INDEX, array_merge($flashes, $this->flashed));
         }
 
         foreach ($this->removed as $type) {
@@ -136,21 +165,13 @@ class CacheSessionService implements SessionService
             $this->storage->set($this->storageKeyFromType($type), $object);
         }
 
-        $this->storage->set(self::FLASHES_STORAGE_INDEX, $this->flashed);
+        $this->storage->set(self::SESSION_EXPIRATION_KEY . $this->session_id, $this->time() + $this->session_ttl);
 
         $this->read_cache = [];
         $this->write_cache = [];
         $this->cleared = false;
 
-        // TODO move this to a session cookie component with an interface?!
-        $cookie_string = sprintf(
-            self::SESSION_COOKIE_KEY . "=%s; Max-Age=%s; Expires=%s; Path=/;",
-            $this->session_id,
-            $this->session_ttl,
-            time() + $this->session_ttl
-        );
-
-        $response = $response->withAddedHeader("set-cookie", $cookie_string);
+        $response = $this->addSessionCookie($response);
 
         return $response;
     }
@@ -177,6 +198,18 @@ class CacheSessionService implements SessionService
         }
     }
 
+    /**
+     * Checks if an object with the class name $type APPEARS to be in cache.
+     *
+     * What is meant by "APPEARS to be in cache"?:
+     * If during the lifetime of the CacheSessionService, the $type was either removed, or the whole session storage
+     * cleared by calling clear(), the connection to the storage is effectively cut-off, so the result reflects whether
+     * the session was cleared or the $type removed since the last commit to storage.
+     *
+     * @param string $type
+     *
+     * @return bool returns true if an object of the given $type appears to be stored in storage.
+     */
     protected function existsInStorage($type)
     {
         if ($this->cleared || isset($this->removed[$type])) {
@@ -189,6 +222,18 @@ class CacheSessionService implements SessionService
 
     }
 
+    /**
+     * If an object of type $type APPEARS to be in cache, then register it in the read_cache array attribute and
+     * return the object.
+     *
+     * What is meant by "APPEARS to be in cache"?:
+     *
+     * @see CacheSessionService::existsInStorage
+     *
+     * @param string $type
+     *
+     * @return mixed|null
+     */
     protected function getFromStorage($type)
     {
         if ($this->cleared || isset($this->removed[$type])) {
@@ -206,8 +251,70 @@ class CacheSessionService implements SessionService
         return $object;
     }
 
+    /**
+     * Prefixes the type name with a session "namespace" and the session_id, so the index in the cache storage is
+     * distinguishable from an index for the same type but a different session.
+     *
+     * @param string $type
+     *
+     * @return string
+     */
     protected function storageKeyFromType($type)
     {
         return self::SESSION_INDEX_PREFIX . $this->session_id . ".$type";
+    }
+
+    /**
+     * @param ResponseInterface $response
+     *
+     * @return ResponseInterface
+     */
+    protected function addSessionCookie(ResponseInterface $response)
+    {
+        $cookie_string = sprintf(
+            self::COOKIE_KEY . "=%s; Max-Age=%s; Expires=%s; Path=/;",
+            $this->session_id,
+            $this->session_ttl,
+            $this->time() + $this->session_ttl
+        );
+
+        $response = $response->withAddedHeader("set-cookie", $cookie_string);
+
+        return $response;
+    }
+
+    /**
+     * @param string|null $session_id
+     *
+     * @return bool
+     */
+    protected function sessionIsActive($session_id)
+    {
+        if ($session_id === null) {
+            return false;
+        }
+
+        $expiration_time = $this->storage->get(self::SESSION_EXPIRATION_KEY . $this->session_id);
+
+        if (is_int($expiration_time) && $expiration_time > $this->time()) {
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * Returns the current value of time().
+     *
+     * The purpose of calling this method instead of the time() function directly is for testing purposes.
+     *
+     * If you want to make a test for this class that is dependent on knowing the current time, make an extension class,
+     * with this method overridden with a known time value, that can be tested against, and test that extension instead.
+     *
+     * @return int the current time in seconds since the Epoch
+     */
+    protected function time()
+    {
+        return time();
     }
 }
